requirements.txt
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.5
duckdb==1.1.3
pandas==2.2.3
python-dotenv==1.0.1
pydantic==2.10.2
typing-extensions==4.12.2 